App building again after moving state up in conundrum parser and starting to implement plaintext and markdown output.

+++ packages/rust/fluster_core_utilities/src/core_types/webviews/note_detail_state/note_detail_state_model.rs
@@ -22,7 +22,7 @@ pub enum SummaryGenerationMethod {
 pub struct SummaryState {
     pub content: String,
     /// The javascript/unix timestamp in milliseconds.
-    pub ctime: u32,
+    pub ctime: f32,
     pub generation_method: SummaryGenerationMethod,
 }
